+++ pkg/routes/api/v1/admin/user_create.go
@@ -20,27 +20,13 @@ import (
 	"errors"
 	"net/http"
 
-	"code.vikunja.io/api/pkg/config"
 	"code.vikunja.io/api/pkg/db"
 	"code.vikunja.io/api/pkg/models"
 	"code.vikunja.io/api/pkg/modules/auth/openid"
-	"code.vikunja.io/api/pkg/user"
+	"code.vikunja.io/api/pkg/routes/api/shared"
 
 	"github.com/labstack/echo/v5"
 )
-
-// CreateUserBody wraps user.APIUserPassword with admin-only fields.
-type CreateUserBody struct {
-	// The full name of the new user. Optional.
-	Name string `json:"name"`
-	// The language of the new user. Must be a valid IETF BCP 47 language code and exist in Vikunja.
-	Language string `json:"language" valid:"language"`
-	user.APIUserPassword
-	// Mark the new user as an instance admin.
-	IsAdmin bool `json:"is_admin"`
-	// Activate the new user immediately without email confirmation.
-	SkipEmailConfirm bool `json:"skip_email_confirm"`
-}
 
 // CreateUser provisions a new account on behalf of an instance admin.
 // @Summary Create a user (admin)
@@ -49,12 +35,12 @@ type CreateUserBody struct {
 // @Accept json
 // @Produce json
 // @Security JWTKeyAuth
-// @Param body body admin.CreateUserBody true "The user to create"
-// @Success 200 {object} admin.User
+// @Param body body models.CreateUserBody true "The user to create"
+// @Success 200 {object} shared.AdminUser
 // @Failure 400 {object} web.HTTPError
 // @Router /admin/users [post]
 func CreateUser(c *echo.Context) error {
-	body := &CreateUserBody{}
+	body := &models.CreateUserBody{}
 	if err := c.Bind(body); err != nil {
 		return c.JSON(http.StatusBadRequest, models.Message{Message: "No or invalid user model provided."})
 	}
@@ -69,46 +55,9 @@ func CreateUser(c *echo.Context) error {
 	s := db.NewSession()
 	defer s.Close()
 
-	newUser, err := models.RegisterUser(s, &user.User{
-		Username: body.Username,
-		Password: body.Password,
-		Email:    body.Email,
-		Name:     body.Name,
-		Language: body.Language,
-	})
+	newUser, err := models.CreateUserAsAdmin(s, body)
 	if err != nil {
 		_ = s.Rollback()
-		return err
-	}
-
-	if body.IsAdmin {
-		if _, err := s.ID(newUser.ID).Cols("is_admin").Update(&user.User{IsAdmin: true}); err != nil {
-			_ = s.Rollback()
-			return err
-		}
-		newUser.IsAdmin = true
-	}
-
-	// Force Active when the admin asked to skip, or when no mailer exists to send the confirmation.
-	if body.SkipEmailConfirm || !config.MailerEnabled.GetBool() {
-		if err := user.SetUserStatus(s, newUser, user.StatusActive); err != nil {
-			_ = s.Rollback()
-			return err
-		}
-		newUser.Status = user.StatusActive
-	}
-
-	if err := s.Commit(); err != nil {
-		_ = s.Rollback()
-		return err
-	}
-
-	// Reload the user so the returned status reflects what was actually persisted
-	// (e.g. StatusEmailConfirmationRequired on mail-enabled instances).
-	rs := db.NewSession()
-	defer rs.Close()
-	newUser, err = user.GetUserByID(rs, newUser.ID)
-	if err != nil {
 		return err
 	}
 
@@ -116,5 +65,5 @@ func CreateUser(c *echo.Context) error {
 	if err != nil {
 		return err
 	}
-	return c.JSON(http.StatusOK, newAdminUser(newUser, providers))
+	return c.JSON(http.StatusOK, shared.NewAdminUser(newUser, providers))
 }
