+++ pkg/routes/api/v1/admin/users_mgmt.go
@@ -23,7 +23,9 @@ import (
 	"code.vikunja.io/api/pkg/db"
 	"code.vikunja.io/api/pkg/models"
 	"code.vikunja.io/api/pkg/modules/auth/openid"
+	"code.vikunja.io/api/pkg/routes/api/shared"
 	"code.vikunja.io/api/pkg/user"
+
 	"github.com/labstack/echo/v5"
 )
 
@@ -41,7 +43,7 @@ type StatusPatch struct {
 // @Security JWTKeyAuth
 // @Param id path int true "User ID"
 // @Param body body admin.StatusPatch true "Status"
-// @Success 200 {object} admin.User
+// @Success 200 {object} shared.AdminUser
 // @Failure 400 {object} web.HTTPError
 // @Failure 404 {object} web.HTTPError
 // @Router /admin/users/{id}/status [patch]
@@ -96,7 +98,7 @@ func PatchStatus(c *echo.Context) error {
 	if err != nil {
 		return err
 	}
-	return c.JSON(http.StatusOK, newAdminUser(target, providers))
+	return c.JSON(http.StatusOK, shared.NewAdminUser(target, providers))
 }
 
 // DeleteUser removes a user either immediately or through the self-deletion flow.
